IPD-KIR Release 2.7.0
The KIR.dat files has been rebuilt to improve the annotation of
sequence features, in particular exon/intron borders. The new method
also includes marking up exons as ‘pseudo’ if the sequence matches a
known exon sequence but is not included in the CDS, i.e. a pseudo exon.
The file formatted is documented in the new Manual.md file. This has
also necessitated changes to the README.md and md5checksum.txt files.